project options system upgrade
